requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
decorator==4.4.1
importlib-metadata==0.23
mock==3.0.5
more-itertools==7.2.0
networkx==2.4
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyaml==19.4.1
pyparsing==2.4.5
pytest==5.2.3
PyYAML==5.1.2
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
